Move job into info.py
